+++ src/main/java/net/geforcemods/securitycraft/blockentities/KeycardReaderBlockEntity.java
@@ -126,7 +126,7 @@ public class KeycardReaderBlockEntity extends DisguisableBlockEntity implements 
 			if (chance < 0.0D)
 				PlayerUtils.sendMessageToPlayer(player, Utils.localize(getBlockState().getBlock().getDescriptionId()), Utils.localize("messages.securitycraft:codebreakerDisabled"), TextFormatting.RED);
 			else {
-				if (isOwnedBy(player) || CodebreakerItem.wasRecentlyUsed(stack))
+				if (!player.isCreative() && (isOwnedBy(player) || CodebreakerItem.wasRecentlyUsed(stack)))
 					return ActionResultType.PASS;
 
 				boolean isSuccessful = player.isCreative() || SecurityCraft.RANDOM.nextDouble() < chance;